all mappings now belong to same controller
